+++ src/mlpack/bindings/python/print_doc_functions_impl.hpp
@@ -37,6 +37,14 @@ inline std::string PrintImport(const std::string& bindingName)
 }
 
 /**
+ * Print any special information about input options.
+ */
+inline std::string PrintInputOptionInfo()
+{
+  return "";
+}
+
+/**
  * Print any special information about output options.
  */
 inline std::string PrintOutputOptionInfo()