+++ HISTORY.md
@@ -1,5 +1,12 @@
 ### mlpack ?.?.?
 ###### ????-??-??
+  * Added Noisy DQN to q_networks (#2446).
+
+  * Add Go bindings (#1884).
+
+  * Added Dueling DQN to q_networks, Noisy linear layer to ann/layer
+    and Empty loss to ann/loss_functions (#2414).
+
   * Storing and adding accessor method for action in q_learning (#2413).
 
   * Added accessor methods for ANN layers (#2321).
@@ -12,10 +19,22 @@
 
   * Add `MatType` parameter to `LSHSearch`, allowing sparse matrices to be used
     for search (#2395).
-    
+
   * Documentation fixes to resolve Doxygen warnings and issues (#2400).
-  
+
   * Add Load and Save of Sparse Matrix (#2344).
+
+  * Add Intersection over Union (IoU) metric for bounding boxes (#2402).
+
+  * Add Non Maximal Supression (NMS) metric for bounding boxes (#2410).
+
+  * Fix `no_intercept` and probability computation for linear SVM bindings
+    (#2419).
+
+  * Fix incorrect neighbors for `k > 1` searches in `approx_kfn` binding, for
+    the `QDAFN` algorithm (#2448).
+
+  * Add `RBF` layer in ann module to make `RBFN` architecture (#2261).
 
 ### mlpack 3.3.1
 ###### 2020-04-29